// src/app/models/car-model.ts
export interface Car {

  id: number;
  brand: string;
  model: string;
  year: number;
  km: number;
  price: number;
  fuelType: string;
  previousOwners: number;
  photoUrl: string;
  transmission: string;
  horsePower: number;
  gearbox: string;
  cylinders: number;
  engineCapacity: number;
  color: string;
  seatMaterial: string;
  description: string;
  location: string;

}